Hash-to-value maps are generated for WPDB, and ...
... the RMs generate SQL builder tokens and value lists for WPDB.

The hash keys generated for the values ensure that duplicate values get different hashes.
This is achieved by using their position as part of the hash.

The RMs create a value-to-token map from the hash-to-value map.
Since the values are now keys, duplicate values get "merged".
This is ok, since the tokens are only used for replacement by the query builder traits.

Finally, the RMs pass a numeric array containing only the values to WPDB.
This list is generated from the hash-to-value map, which processes values in order.
Therfore, thee order of the values should be preserved.